Add go program to populate the ASN mappings table
This program uses the ASRank API to download and insert all of
the ASNs and AS Names in that data set.

The "Changed" field in each ASN returned is interpreted as the
"apply from" timestamp in our mapping database.

Ideally, you'll want to run this program fairly regularly to
capture any changes / additions to the ASN set.