Update ASN label scraper to match new structure of AS info
The latest AS2Org update has added more detail to the OrgId
field returned for each AS, so we now parse that correctly.

Included in that detail is the Org Name, so I've also added
this as a column to include in our SQLite database as this
name might be preferable to use as a label rather than the
AS name (which can be rather cryptic).